+++ src/ts/popup-closing.ts
@@ -1,8 +1,0 @@
-document.addEventListener('DOMContentLoaded', () => {
-
-  if (window.opener) {
-    const redirect = '' !== window.location.search ? '&order_payed' : '?order_payed';
-    window.opener.location = window.location.href + redirect;
-    window.close();
-  }
-});